Moved `WavelengthCalibration._cross_correlation` to `core` as `cross_correlation`. Also moved test, though it is been skipped and continued that way.
