+++ goodman_pipeline/core/__init__.py
@@ -21,6 +21,7 @@ from .core import (astroscrappy_lacosmic,
                    convert_time,
                    create_master_bias,
                    create_master_flats,
+                   cross_correlation,
                    dcr_cosmicray_rejection,
                    define_trim_section,
                    extraction,
